+++ server/standalone/migration/000023_add_algo_input_output.down.sql
@@ -1,0 +1,7 @@
+SELECT execute($$
+
+    DROP TABLE algo_inputs;
+    DROP TABLE algo_outputs;
+    DROP TABLE asset_kinds;
+
+$$) WHERE table_exists('public', 'algo_inputs');