+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Remove performance asset column (#640)
 - Remove datamanager asset column (#652)
 - Remove datasample asset column (#666)
+- Algos now have Inputs and Outputs (#641)
 
 ## [0.9.2] - 2022-04-15